Changes in sending service

+++ app/src/main/java/com/example/jorge/blue/servicios/SendingService.java
@@ -2,29 +2,41 @@ package com.example.jorge.blue.servicios;
 
 import android.annotation.SuppressLint;
 import android.app.Service;
+import android.content.Context;
 import android.content.Intent;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
 import android.os.Binder;
 import android.os.IBinder;
 import android.os.PowerManager;
 import android.util.Log;
+
+import com.example.jorge.blue.activities.UserInterfaz;
 import com.example.jorge.blue.entidades.ConexionSQLiteHelper;
 import com.example.jorge.blue.utils.Identifiers;
 import com.example.jorge.blue.utils.Utilities;
+
+import okhttp3.MediaType;
 import okhttp3.OkHttpClient;
+import okhttp3.Request;
+import okhttp3.RequestBody;
+import okhttp3.Response;
+
 import org.json.JSONArray;
 import org.json.JSONObject;
 import java.io.DataOutputStream;
+import java.io.IOException;
 import java.net.HttpURLConnection;
 import java.net.URL;
+import java.util.concurrent.TimeUnit;
 
 public class SendingService extends Service {
     private final IBinder mBinder = new LocalBinder();
     private final String TAG = "SENDING SERVICE";
     private boolean responseId;
     private OkHttpClient okHttpClient;
-    private ConexionSQLiteHelper connection;
     public static PowerManager.WakeLock wakeLock;
     public static Thread thread;
 
@@ -42,10 +54,31 @@ public class SendingService extends Service {
     @SuppressLint("InvalidWakeLockTag")
     @Override
     public void onCreate() {
-        responseId = Utilities.getStationID(okHttpClient);
+        boolean res = false;
+
+        OkHttpClient.Builder builder = new OkHttpClient.Builder();
+        builder.connectTimeout(300, TimeUnit.SECONDS);
+        builder.readTimeout(300, TimeUnit.SECONDS);
+        builder.writeTimeout(300, TimeUnit.SECONDS);
+        okHttpClient = builder.build();
+
+        Identifiers.connection = new ConexionSQLiteHelper(this, "medicion", null, 1);
+        while(!res) {
+            ConnectivityManager cm = (ConnectivityManager) getApplicationContext().getSystemService(Context.CONNECTIVITY_SERVICE);
+            NetworkInfo activeNetwork = cm.getActiveNetworkInfo();
+            if (activeNetwork != null && activeNetwork.isConnected()) {
+                responseId = Utilities.getStationID(okHttpClient);
+                res = true;
+            } else {
+                try {
+                    Thread.sleep(60000);
+                } catch (InterruptedException e) {
+                    Log.e(TAG, "NO HAY CONEXIÓN A INTERNET. SE VOLVERÁ A INTENTAR EN 60 SEGUNDOS");
+                    e.printStackTrace();
+                }
+            }
+        }
         Identifiers.setAPIKey(getApplicationContext());
-        connection = new ConexionSQLiteHelper(getApplicationContext(), "medicion", null, 1);
-        okHttpClient = new OkHttpClient();
 
         //MANTENER ENCENDIDO EL CPU DEL CELULAR AL APAGAR LA PANTALLA
         PowerManager powerManager = (PowerManager) getSystemService(POWER_SERVICE);
@@ -56,7 +89,7 @@ public class SendingService extends Service {
     @Override
     public int onStartCommand(Intent intent, int flags, int startId) {
         responseId = Utilities.getStationID(okHttpClient);
-        //sendPost();
+        sendPost();
         return START_STICKY;
     }
 
@@ -70,36 +103,60 @@ public class SendingService extends Service {
         thread = new Thread(new Runnable() {
             @Override
             public void run() {
-                try {
-                    URL url = new URL(Identifiers.URL_SERVER);
-                    HttpURLConnection connect = (HttpURLConnection) url.openConnection();
-                    connect.setRequestMethod("POST");
-                    connect.setRequestProperty("Content-Type", "application/json;charset=UTF-8");
-                    connect.setRequestProperty("Accept","application/json");
-                    connect.setDoOutput(true);
-                    connect.setDoInput(true);
+                boolean res = false;
+                while(!res) {
+                    try {
+                        Log.i(TAG, "ENTRO");
 
-                    JSONObject jsonParam = consultarMediciones();
-                    Log.i(TAG, "DATOS A ENVIAR AL SERVIDOR: " + jsonParam.toString());
+                        OkHttpClient.Builder builder = new OkHttpClient.Builder();
+                        builder.connectTimeout(300, TimeUnit.SECONDS);
+                        builder.readTimeout(300, TimeUnit.SECONDS);
+                        builder.writeTimeout(300, TimeUnit.SECONDS);
+                        okHttpClient = builder.build();
 
-                    DataOutputStream os = new DataOutputStream(connect.getOutputStream());
-                    os.writeBytes(jsonParam.toString());
-                    os.flush();
-                    os.close();
+                        MediaType JSON = MediaType.parse("application/json; charset=utf-8");
+                        JSONObject jsonParam = consultarMediciones();
+                        Log.i(TAG, "DATOS A ENVIAR AL SERVIDOR: " + jsonParam.toString());
+                        RequestBody body = RequestBody.create(JSON, String.valueOf(jsonParam));
+                        Request request = new Request.Builder()
+                                .url(Identifiers.URL_SERVER)
+                                .post(body)
+                                .build();
+                        Identifiers.callSending = okHttpClient.newCall(request);
+                        borrarBD();
+                        Log.i(TAG, "DATOS BORRADOS Y ENVIÁNDOSE");
+                        Identifiers.callSending.execute();
+                        /*int respuesta = response.code();
+                        Log.i(TAG, "CÓDIGO DE RESPUESTA DEL SERVIDOR: " + String.valueOf(respuesta));*/
 
-                    int st = connect.getResponseCode();
-                    Log.d(TAG, "CÓDIGO DE RESPUESTA DEL SERVIDOR: " + String.valueOf(st));
-                    //Log.d("MSG" , connect.getResponseMessage());
+                        /*URL url = new URL(Identifiers.URL_SERVER);
+                        HttpURLConnection connect = (HttpURLConnection) url.openConnection();
+                        connect.setRequestMethod("POST");
+                        connect.setRequestProperty("Content-Type", "application/json;charset=UTF-8");
+                        connect.setRequestProperty("Accept", "application/json");
+                        connect.setDoOutput(true);
+                        connect.setDoInput(true);
 
-                    connect.disconnect();
-                    if (st == 200) {
+                        JSONObject jsonParam = consultarMediciones();
+                        Log.i(TAG, "DATOS A ENVIAR AL SERVIDOR: " + jsonParam.toString());
                         borrarBD();
                         Log.i(TAG, "DATOS ENVIADOS Y BORRADOS");
+
+                        DataOutputStream os = new DataOutputStream(connect.getOutputStream());
+                        os.writeBytes(jsonParam.toString());
+                        os.flush();
+                        os.close();
+
+                        int st = connect.getResponseCode();
+                        Log.i(TAG, "CÓDIGO DE RESPUESTA DEL SERVIDOR: " + String.valueOf(st));
+
+                        connect.disconnect();*/
+                        Thread.sleep(5000);
+                    } catch(Exception e) {
+                        Log.e(TAG, "ERROR AL ENVIAR LOS DATOS: " + e.getMessage());
+                        e.printStackTrace();
+                        res = true;
                     }
-                    //Thread.sleep(10000);
-                } catch (Exception e) {
-                    Log.e(TAG, "ERROR DE ENVÍO DE DATOS: " + e.getMessage());
-                    e.printStackTrace();
                 }
             }
         });
@@ -108,40 +165,64 @@ public class SendingService extends Service {
     }
 
     public JSONObject consultarMediciones() {
-        SQLiteDatabase db = connection.getReadableDatabase();
-        Cursor cursor = db.rawQuery("SELECT * FROM " + Identifiers.TABLA_MEDICION, null);
-        db.close();
-        JSONArray jsonArray = new JSONArray();
-        JSONObject y = new JSONObject();
-        if(responseId) {
-            try {
-                while (cursor.moveToNext()) {
-                    JSONObject j = new JSONObject();
-                    j.put("StationId", Identifiers.ID_STATION);
-                    j.put("Timestamp", cursor.getString(0));
-                    j.put("Type", cursor.getString(1));
-                    j.put("Value", cursor.getString(2));
-                    j.put("Units", cursor.getString(3));
-                    j.put("Location", cursor.getString(5));
-                    j.put("SensorId", cursor.getString(4));
-                    jsonArray.put(j);
+        SQLiteDatabase db = Identifiers.connection.getReadableDatabase();
+        /*while(true) {
+            Log.e(TAG, "LA BASE ESTÁ ABIERTA 1: " + db.isOpen());
+            if(!db.isOpen()) {*/
+                Cursor cursor = db.rawQuery("SELECT * FROM " + Identifiers.TABLA_MEDICION, null);
+                JSONArray jsonArray = new JSONArray();
+                JSONObject y = new JSONObject();
+                if(responseId) {
+                    try {
+                        while(cursor.moveToNext()) {
+                            JSONObject j = new JSONObject();
+                            j.put("StationId", Identifiers.ID_STATION);
+                            j.put("Timestamp", cursor.getString(0));
+                            j.put("Type", cursor.getString(1));
+                            j.put("Value", cursor.getString(2));
+                            j.put("Units", cursor.getString(3));
+                            j.put("Location", cursor.getString(5));
+                            j.put("SensorId", cursor.getString(4));
+                            jsonArray.put(j);
+                        }
+                        y.put("data", jsonArray);
+                    } catch (Exception e) {
+                        Log.e(TAG, "ERROR AL CARGAR DATOS DE LA BASE DEL DISPOSITIVO: " + e.getMessage());
+                        e.printStackTrace();
+                    }
+                    db.close();
+                    Identifiers.connection.close();
+                    return y;
                 }
-                y.put("data", jsonArray);
-            } catch (Exception e) {
-                Log.e(TAG, "ERROR AL CARGAR DATOS DE LA BASE DEL DISPOSITIV: " + e.getMessage());
+                /*break;
+            }
+            try {
+                Thread.sleep(5000);
+            } catch (InterruptedException e) {
+                Log.e(TAG, "ERROR AL GUARDAR, LA BASE ESTÁ OCUPADA");
                 e.printStackTrace();
             }
-            connection.close();
-            return y;
-        }
+        }*/
         return null;
     }
 
     public void borrarBD() {
-        SQLiteDatabase db = connection.getReadableDatabase();
-        db.execSQL("DELETE FROM " + Identifiers.TABLA_MEDICION);
-        connection.close();
-        Log.i(TAG, "DATOS BORRADOS DE LA BASE: " + Identifiers.TABLA_MEDICION);
+        SQLiteDatabase db = Identifiers.connection.getReadableDatabase();
+        /*while(true) {
+            Log.e(TAG, "LA BASE ESTÁ ABIERTA 2: " + db.isOpen());
+            if (!db.isOpen()) {*/
+                db.execSQL("DELETE FROM " + Identifiers.TABLA_MEDICION);
+                Identifiers.connection.close();
+                Log.i(TAG, "DATOS BORRADOS DE LA BASE: " + Identifiers.TABLA_MEDICION);
+                /*break;
+            }
+            try {
+                Thread.sleep(5000);
+            } catch (InterruptedException e) {
+                Log.e(TAG, "ERROR AL GUARDAR, LA BASE ESTÁ OCUPADA");
+                e.printStackTrace();
+            }
+        }*/
     }
 
 }
